Feat: Enhance SEO and metadata
Update site metadata, add favicons and open graph images to improve SEO and user experience. This change provides richer previews when the site is shared on social media and better identifies the website in search results. It also adds autodeploy config to automatically deploy to production when the main branch is pushed.