Feat: Add X and Discord icons to footer
Adds the X (Twitter) and Discord icons to the footer, linking to the respective Forge platform accounts. This enhances community engagement and makes it easier for users to connect with the platform on different social media channels.

Also removes the `url` from the `ForgePlatform` in the `sst-env.d.ts` file.